EvalAct changes to reflect Categories; UI changes
Commit:

-Evaluations Activity changed so that evaluations shown reflect the
category listview item that sent the intent. Sorting works with this new
feature as well (so sorting by date will not bring up other category's
evaluations, but only if this is not the "all" category) Also, category
filtering option will not show up when a non "all" category item sent
the intent

-Changed UI to make it more mellow and smoother.

One weird thing: groupIndicator icon is weird; i removed the
groupindicator icon code in the XML but now there seems to be a default
one when you launch the app, even though the code does not call for one
O_o Anojh Gnanachandran any ideas?