Refactored classes, idGet_Mode uses Constants
The way I decided to have the packages was inspired by the way it's done
at my workplace.

Rather than using an enum (will be too verbose, or we'd have to pass in
variables of type "EditMode"), or an interface (arguably an
antipattern), I used a public final class with a private constructor
(can't be instantiated or subclassed) for the mode constants (add = 0,
edit = 1). This is better than relying on documentation and 'ints with
special values'.

Refactoring also made me change the Manifest (which still needs to be
changed); there may be some bugs.